requirements/base.txt
# Trading & Binance
python-binance==1.0.32  # Official Binance Python SDK

# Logging
sentry-sdk==2.45.0

# AI and ML packages
llama-index==0.14.8
llama-index-tools-mcp==0.4.3
llama-index-llms-bedrock-converse==0.12.1
json-repair==0.54.2
langfuse==3.10.1
openinference-instrumentation-llama-index==4.3.9
ipython==9.7.0
pandas==2.2.3
ta==0.11.0
llama-index-utils-workflow==0.5.1
apscheduler==3.11.1 # Background task scheduler
nest_asyncio==1.6.0

# Django
Django==5.2.7
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-filter==25.2
django-utils-six==2.0
django-extensions==4.1
django-cors-headers==4.9.0
django-easy-health-check==1.0.1 # https://pypi.org/project/django-easy-health-check/
adrf==0.1.9

# Django rest framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0  # OpenAPI 3.0 documentation

# Services
whitenoise==6.11.0
gunicorn==22.0.0

# Storage
django-storages==1.14.6
psycopg[binary]==3.2.10